Seperate member to get, post, put model
